+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Change Log for SD.Next
 
-## Update for 2025-03-28
+## Update for 2025-03-30
 
-### Highlights for 2025-03-28
+### Highlights for 2025-03-30
 
 Brand new Video processing module with support for all latest models: **WAN21, Hunyuan, LTX, Cog, Allegro, Mochi1, Latte1** in both T2V and I2V workflows  
 And combined with *on-the-fly quantization*, support for *Local/Tiny/Remote* VAE, acceleration modules such as *FasterCache or PAB* and more!  
@@ -15,7 +15,7 @@ Plus...
 - More quantization options and granular control  
 - Pretty big performance updates to a) Any model using DiT based architecture: new caching methods, b) ZLUDA: new attention methods  
 
-### Details for 2025-03-28
+### Details for 2025-03-30
 
 - **Video tab**  
   - see [Video Wiki](https://github.com/vladmandic/sdnext/wiki/Video) for details!  
@@ -129,6 +129,10 @@ Plus...
   - **LoRA** enable memory cache by default  
   - **Samplers** add ability to set sigma adjustment for each sampler  
   - **ModernUI** updates  
+  - **CSS** updates  
+  - settings vertiocal/dirty indicator restores to default setting instead to previous value  
+  - video interpolate do not skip duplicate frames  
+  - **settings UI** full refactor  
 - **Wiki/Docs**  
   - updated [Models](https://github.com/vladmandic/sdnext/wiki/Models) info  
   - new [Video](https://github.com/vladmandic/sdnext/wiki/Video) guide